Made it Responsive
