+++ src/Components/Constants/constants.jsx
@@ -1,0 +1,1 @@
+export {default as archit} from '../images/archit.jpeg';
