+++ src/Components/Events/Event1.jsx
@@ -6,6 +6,7 @@ import { Swiper, SwiperSlide } from "swiper/react";
 // Import Swiper styles
 import "swiper/css";
 import "swiper/css/effect-fade";
+import './Events.css';
 // import required modules
 import { EffectFade, Autoplay } from "swiper";
 
@@ -13,32 +14,21 @@ import { SectionHeading } from "../Global/GlobalStyles";
 
 const Event1 = () => {
   return (
-    <div id="event1">
-      <Container>
-        <Grid container>
+      <Container maxWidth="md" >
+        <Grid container sx={{padding:{lg:'0px 100px 20px 70px'}}}
+          >
           <Grid item xs={12} sm={12} md={12} lg={12} xl={12}>
             <SectionHeading
               variant="h4"
-              sx={{ marginTop: "30px", marginBottom: "50px" }}
+              sx={{ marginTop: "20px", marginBottom: "20px" }}
             >
               Event 1
             </SectionHeading>
           </Grid>
-          <Grid
-            item
-            xs={12}
-            sm={12}
-            md={12}
-            lg={12}
-            xl={12}
-            sx={{
-              display: "flex",
-              flexDirection: "row",
-              justifyContent: "center",
-              alignItems: "center",
-              flexWrap: "wrap",
-            }}
-          >
+        </Grid>
+        <Grid item 
+          
+        >
             <Swiper
               spaceBetween={30}
               centeredSlides={true}
@@ -64,10 +54,10 @@ const Event1 = () => {
                   }}
                 ><SwiperSlide>
                   <img
+                  class="imgevents"
                     key={item.img}
                     style={{
                       borderRadius: "40px",
-                      width: "500px",
                       objectFit: "crop",
                       auto: "format",
                     }}
@@ -81,10 +71,8 @@ const Event1 = () => {
                 </Container>
               ))}
             </Swiper>
-          </Grid>
         </Grid>
       </Container>
-    </div>
   );
 };
