+++ src/Components/Events/Event4.jsx
@@ -2,7 +2,9 @@
 import React from "react";
 import { Container, Grid } from "@mui/material";
 
+
 import { Swiper, SwiperSlide } from "swiper/react";
+import './Events.css';
 
 // Import Swiper styles
 import "swiper/css";
@@ -14,32 +16,21 @@ import { SectionHeading } from "../Global/GlobalStyles";
 
 const Event4 = () => {
   return (
-    <div id="event1">
-      <Container>
-        <Grid container>
+      <Container maxWidth="md" >
+        <Grid container sx={{padding:{lg:'0px 100px 20px 70px'}}}
+          >
           <Grid item xs={12} sm={12} md={12} lg={12} xl={12}>
             <SectionHeading
               variant="h4"
-              sx={{ marginTop: "30px", marginBottom: "50px" }}
+              sx={{ marginTop: "50px", marginBottom: "20px" }}
             >
               Event 4
             </SectionHeading>
           </Grid>
-          <Grid
-            item
-            xs={12}
-            sm={12}
-            md={12}
-            lg={12}
-            xl={12}
-            sx={{
-              display: "flex",
-              flexDirection: "row",
-              justifyContent: "center",
-              alignItems: "center",
-              flexWrap: "wrap",
-            }}
-          >
+        </Grid>
+        <Grid item 
+          
+        >
             <Swiper
               spaceBetween={30}
               centeredSlides={true}
@@ -65,10 +56,10 @@ const Event4 = () => {
                   }}
                 ><SwiperSlide>
                   <img
+                  class="imgevents"
                     key={item.img}
                     style={{
                       borderRadius: "40px",
-                      width: "500px",
                       objectFit: "crop",
                       auto: "format",
                     }}
@@ -82,10 +73,8 @@ const Event4 = () => {
                 </Container>
               ))}
             </Swiper>
-          </Grid>
         </Grid>
       </Container>
-    </div>
   );
 };
