+++ src/Components/Events/EventStyles.jsx
@@ -4,11 +4,11 @@ import { styled } from "@mui/material/styles";
 export const EventContainer = styled(Grid)({
     display:'flex',
     flexDirection:'row',
-    justifyContent:'space-around',
+    justifyContent:'center',
     alignItems:'center',
     flexWrap:'wrap',
 })
 
 export const EventItem = styled(Grid)({
-    maxHeight:'450px'
+    maxHeight:'600px'
 })
