+++ src/Components/Events/Events.css
@@ -1,0 +1,34 @@
+@media  screen and (max-width : 300px){
+    .imgevents{
+        width: 200px;
+    }
+}
+@media  (min-width : 301px) and (max-width : 600px){
+    .imgevents{
+        width: 350px;
+    }
+}
+
+@media  (min-width : 601px) and (max-width : 900px){
+    .imgevents{
+        width: 500px;
+    }
+}
+
+@media  (min-width : 901px) and (max-width : 1200px){
+    .imgevents{
+        width: 600px;
+    }
+}
+
+@media  (min-width : 1201px) and (max-width : 1500px){
+    .imgevents{
+        width: 750px;
+    }
+}
+
+@media  (min-width : 1501px) and (max-width : 1800px){
+    .imgevents{
+        width: 850px;
+    }
+}
