+++ src/Components/Photos/Photos.css
@@ -1,0 +1,34 @@
+@media  screen and (max-width : 300px){
+    #imgphotos{
+        width: 200px;
+    }
+}
+@media  (min-width : 301px) and (max-width : 600px){
+    #imgphotos{
+        width: 350px;
+    }
+}
+
+@media  (min-width : 601px) and (max-width : 900px){
+    #imgphotos{
+        width: 500px;
+    }
+}
+
+@media  (min-width : 901px) and (max-width : 1200px){
+    #imgphotos{
+        width: 600px;
+    }
+}
+
+@media  (min-width : 1201px) and (max-width : 1500px){
+    #imgphotos{
+        width: 750px;
+    }
+}
+
+@media  (min-width : 1501px) and (max-width : 1800px){
+    #imgphotos{
+        width: 850px;
+    }
+}
