+++ src/Components/Team/Teamm.js
@@ -1,0 +1,116 @@
+const teamMembers = [
+    {
+      img: "https://images.pexels.com/photos/762020/pexels-photo-762020.jpeg?auto=compress&cs=tinysrgb&w=600",
+      title: "Kylie",
+      position: "Software developer",
+      Facebook: "https://www.facebook.com",
+      Github: "https://www.github.com",
+      Instagram: "https//www.instagram.com",
+      Twitter: "https://www.facebook.com",
+    },
+    {
+      img: "https://images.pexels.com/photos/774909/pexels-photo-774909.jpeg?auto=compress&cs=tinysrgb&w=600",
+      title: "Liza",
+      position: "Front-end Develepor",
+      Facebook: "https://www.facebook.com",
+      Github: "https://www.github.com",
+      Instagram: "https//www.instagram.com",
+      Twitter: "https://www.facebook.com",
+    },
+    {
+      img: "https://images.pexels.com/photos/774095/pexels-photo-774095.jpeg?auto=compress&cs=tinysrgb&w=600",
+      title: "Cheryl",
+      position: "Back-end Developer",
+      Facebook: "https://www.facebook.com",
+      Github: "https://www.github.com",
+      Instagram: "https//www.instagram.com",
+      Twitter: "https://www.facebook.com",
+    },
+    {
+      img: "https://images.pexels.com/photos/1222271/pexels-photo-1222271.jpeg?auto=compress&cs=tinysrgb&w=600",
+      title: "David",
+      position: "Web Develepor",
+      Facebook: "https://www.facebook.com",
+      Github: "https://www.github.com",
+      Instagram: "https//www.instagram.com",
+      Twitter: "https://www.facebook.com",
+    },
+    
+    {
+      img: "https://images.pexels.com/photos/614810/pexels-photo-614810.jpeg?auto=compress&cs=tinysrgb&w=600",
+      title: "Sam",
+      position: "Front-end Develepor",
+      Facebook: "https://www.facebook.com",
+      Github: "https://www.github.com",
+      Instagram: "https//www.instagram.com",
+      Twitter: "https://www.facebook.com",
+    },
+    {
+      img: "https://images.pexels.com/photos/532220/pexels-photo-532220.jpeg?auto=compress&cs=tinysrgb&w=600",
+      title: "Max",
+      position: "Back-end Develepor",
+      Facebook: "https://www.facebook.com",
+      Github: "https://www.github.com",
+      Instagram: "https//www.instagram.com",
+      Twitter: "https://www.facebook.com",
+    },
+    {
+      img: "https://images.pexels.com/photos/1065084/pexels-photo-1065084.jpeg?auto=compress&cs=tinysrgb&w=600",
+      title: "Suzen",
+      position: "Web Develepor",
+      Facebook: "https://www.facebook.com",
+      Github: "https://www.github.com",
+      Instagram: "https//www.instagram.com",
+      Twitter: "https://www.facebook.com",
+    },
+    {
+      img: "https://images.pexels.com/photos/372042/pexels-photo-372042.jpeg?auto=compress&cs=tinysrgb&w=600",
+      title: "Riya",
+      position: "Web Developer",
+      Facebook: "https://www.facebook.com",
+      Github: "https://www.github.com",
+      Instagram: "https//www.instagram.com",
+      Twitter: "https://www.facebook.com",
+    },
+    {
+      img: "https://images.pexels.com/photos/874158/pexels-photo-874158.jpeg?auto=compress&cs=tinysrgb&w=600",
+      title: "Robert",
+      position: "Back-end Develepor",
+      Facebook: "https://www.facebook.com",
+      Github: "https://www.github.com",
+      Instagram: "https//www.instagram.com",
+      Twitter: "https://www.facebook.com",
+    },
+    {
+      img: "https://images.pexels.com/photos/38554/girl-people-landscape-sun-38554.jpeg?auto=compress&cs=tinysrgb&w=600",
+      title: "Megan",
+      position: "Back-end Develepor",
+      Facebook: "https://www.facebook.com",
+      Github: "https://www.github.com",
+      Instagram: "https//www.instagram.com",
+      Twitter: "https://www.facebook.com",
+    },
+    {
+      img: "https://images.pexels.com/photos/1212984/pexels-photo-1212984.jpeg?auto=compress&cs=tinysrgb&w=600",
+      title: "Thomas",
+      position: "Web Develepor",
+      Facebook: "https://www.facebook.com",
+      Github: "https://www.github.com",
+      Instagram: "https//www.instagram.com",
+      Twitter: "https://www.facebook.com",
+    },
+    {
+      img: "https://images.pexels.com/photos/775358/pexels-photo-775358.jpeg?auto=compress&cs=tinysrgb&w=600",
+      title: "John",
+      position: "Front-end Develepor",
+      Facebook: "https://www.facebook.com",
+      Github: "https://www.github.com",
+      Instagram: "https//www.instagram.com",
+      Twitter: "https://www.facebook.com",
+    },
+    
+    
+  ];
+  
+
+  export default teamMembers;
